refactor(ui): Fix fast refresh by only exporting components
If a file exports something else in addition to the component created in
it, fast refresh doesn't work[1]. Move exports from the button and form
components to separate files to allow fast refresh.

[1]: https://github.com/ArnaudBarre/eslint-plugin-react-refresh
