+++ ui/src/components/ui/alert-dialog.tsx
@@ -11,7 +11,7 @@ import * as React from 'react';
 import * as AlertDialogPrimitive from '@radix-ui/react-alert-dialog';
 
 import { cn } from '@/lib/utils';
-import { buttonVariants } from '@/components/ui/button';
+import { buttonVariants } from '@/components/ui/button-variants';
 
 const AlertDialog = AlertDialogPrimitive.Root;
